# Add check MIME type functions to API