Added in API of this repository an opportunity to check whether content matches a specific MIME type. For example this is could be used to detect that file isn't malformed or encrypted by ransomware much more faster (if we have a hint by its extension or content type).

Before this PR:
```go
func checkFileIsOK(f *os.File) (ok bool, err error) {
    ext := path.Ext(f.Name())
    mime, err := mimetype.DetectReader(f)
    if err != nil {
        return false, err
    }
    return mime.Extension() == ext, nil
}
```
Benchmarks:
```
BenchmarkDetectSliceTar-8    	 5507715	       211 ns/op
BenchmarkDetectSliceZip-8    	 2538682	       474 ns/op
BenchmarkDetectSliceJpeg-8   	14020256	        83.1 ns/op
BenchmarkDetectSliceGif-8    	 6674289	       169 ns/op
BenchmarkDetectSlicePng-8    	14576334	        87.2 ns/op
```
After this PR:
```go
func checkFileIsOK(f *os.File) (ok bool, err error) {
    ext := path.Ext(f.Name())
    return mimetype.MatchReaderExtension(f, ext)
}
```
Benchmarks:
```
BenchmarkMatchSliceTar-8     	33561585	        33.4 ns/op
BenchmarkMatchSliceZip-8     	36062886	        30.8 ns/op
BenchmarkMatchSliceJpeg-8    	63828153	        17.7 ns/op
BenchmarkMatchSliceGif-8     	64112824	        18.2 ns/op
BenchmarkMatchSlicePng-8     	65837989	        17.7 ns/op
```
So if we have lots of files to check, it's a huge difference.