@gabriel-vasile 2024-11-11
I can see the idea behind this PR, but until now no one asked for this feature.